--- src/web/middleware/connect/index.ts
@@ -1,4 +1,0 @@
-export * from './connect-asymmetric-jwt-auth-middleware';
-export * from './connect-context-symmetric-jwt-auth-middleware';
-export * from './connect-server-symmetric-jwt-auth-middleware';
-export * from './extract-user-id-from-headers-middleware';
