+++ src/web/middleware/jira/extract-user-id-from-headers-middleware.ts
