+++ src/web/middleware/jira/jira-context-symmetric-jwt-auth-middleware.ts
@@ -5,7 +5,7 @@ import {
 } from 'atlassian-jwt';
 import type { NextFunction, Request, RequestHandler, Response } from 'express';
 
-import { verifyExp, verifyUrlBoundQsh } from './jwt-utils';
+import { verifyExp, verifyFixedQsh, verifyIss } from './jwt-utils';
 import { CONNECT_JWT_TOKEN_CLAIMS_SCHEMA } from './schemas';
 
 import { isEnumValueOf } from '../../../common/enumUtils';
@@ -14,18 +14,20 @@ import { assertSchema, getLogger } from '../../../infrastructure';
 import { connectInstallationRepository } from '../../../infrastructure/repositories';
 import { UnauthorizedError } from '../errors';
 
+const CONTEXT_TOKEN_QSH = 'context-qsh';
+
 /**
- * Authenticates requests using an iframe or server-to-server symmetric JWT token.
+ * Authenticates requests using a context symmetric JWT token.
  *
  * In case of successful authentication, `connectInstallation` and `atlassianUserId` are set in locals.
  *
  * @remarks
- * An iframe or server-to-server symmetric JWT tokens are sent by Jira server.
+ * Context JWT tokens are sent by Connect App UI embed in Jira via extension points.
  *
  * @see https://developer.atlassian.com/cloud/jira/platform/understanding-jwt-for-connect-apps/#types-of-jwt-token
  * @see https://community.developer.atlassian.com/t/action-required-atlassian-connect-vulnerability-allows-bypass-of-app-qsh-verification-via-context-jwts/47072
  */
-export const connectServerSymmetricJwtAuthMiddleware: RequestHandler = (
+export const jiraContextSymmetricJwtAuthMiddleware: RequestHandler = (
 	req: Request,
 	res: Response,
 	next: NextFunction,
@@ -36,23 +38,21 @@ export const connectServerSymmetricJwtAuthMiddleware: RequestHandler = (
 		return next(new UnauthorizedError('Missing JWT token.'));
 	}
 
-	void verifyServerSymmetricJwtToken(req, token)
-		.then(({ connectInstallation }) => {
+	void verifyContextSymmetricJwtToken(token)
+		.then(({ connectInstallation, atlassianUserId }) => {
 			res.locals.connectInstallation = connectInstallation;
+			res.locals.atlassianUserId = atlassianUserId;
 			next();
 		})
-		.catch(next);
+		.then(next);
 };
 
 /**
  * @see https://developer.atlassian.com/cloud/jira/platform/understanding-jwt-for-connect-apps/#decoding-and-verifying-a-jwt-token
  */
-const verifyServerSymmetricJwtToken = async (
-	request: Request,
-	token: string,
-) => {
+const verifyContextSymmetricJwtToken = async (token: string) => {
 	try {
-		const tokenSigningAlgorithm: unknown = getAlgorithm(token);
+		const tokenSigningAlgorithm = getAlgorithm(token) as unknown;
 
 		if (!isEnumValueOf(SymmetricAlgorithm, tokenSigningAlgorithm)) {
 			throw new UnauthorizedError('Unsupported JWT signing algorithm.');
@@ -68,6 +68,8 @@ const verifyServerSymmetricJwtToken = async (
 
 		assertSchema(unverifiedClaims, CONNECT_JWT_TOKEN_CLAIMS_SCHEMA);
 
+		verifyIss(unverifiedClaims);
+
 		const connectInstallation =
 			await connectInstallationRepository.getByClientKey(
 				ensureString(unverifiedClaims.iss),
@@ -81,14 +83,15 @@ const verifyServerSymmetricJwtToken = async (
 
 		assertSchema(verifiedClaims, CONNECT_JWT_TOKEN_CLAIMS_SCHEMA);
 
-		verifyUrlBoundQsh(verifiedClaims, request);
+		verifyFixedQsh(verifiedClaims, CONTEXT_TOKEN_QSH);
 		verifyExp(verifiedClaims);
 
 		return {
 			connectInstallation,
+			atlassianUserId: ensureString(verifiedClaims.sub),
 		};
 	} catch (e) {
-		getLogger().warn(e, 'Failed to verify the server symmetric JWT token.');
+		getLogger().warn(e, 'Failed to verify the context symmetric JWT token.');
 
 		if (e instanceof UnauthorizedError) throw e;
