+++ src/web/middleware/jira/jira-server-symmetric-jwt-auth-middleware.ts
@@ -5,7 +5,7 @@ import {
 } from 'atlassian-jwt';
 import type { NextFunction, Request, RequestHandler, Response } from 'express';
 
-import { verifyExp, verifyFixedQsh } from './jwt-utils';
+import { verifyExp, verifyIss, verifyUrlBoundQsh } from './jwt-utils';
 import { CONNECT_JWT_TOKEN_CLAIMS_SCHEMA } from './schemas';
 
 import { isEnumValueOf } from '../../../common/enumUtils';
@@ -14,20 +14,18 @@ import { assertSchema, getLogger } from '../../../infrastructure';
 import { connectInstallationRepository } from '../../../infrastructure/repositories';
 import { UnauthorizedError } from '../errors';
 
-const CONTEXT_TOKEN_QSH = 'context-qsh';
-
 /**
- * Authenticates requests using a context symmetric JWT token.
+ * Authenticates requests using an iframe or server-to-server symmetric JWT token.
  *
  * In case of successful authentication, `connectInstallation` and `atlassianUserId` are set in locals.
  *
  * @remarks
- * Context JWT tokens are sent by Connect App UI embed in Jira via extension points.
+ * An iframe or server-to-server symmetric JWT tokens are sent by Jira server.
  *
  * @see https://developer.atlassian.com/cloud/jira/platform/understanding-jwt-for-connect-apps/#types-of-jwt-token
  * @see https://community.developer.atlassian.com/t/action-required-atlassian-connect-vulnerability-allows-bypass-of-app-qsh-verification-via-context-jwts/47072
  */
-export const connectContextSymmetricJwtAuthMiddleware: RequestHandler = (
+export const jiraServerSymmetricJwtAuthMiddleware: RequestHandler = (
 	req: Request,
 	res: Response,
 	next: NextFunction,
@@ -38,21 +36,23 @@ export const connectContextSymmetricJwtAuthMiddleware: RequestHandler = (
 		return next(new UnauthorizedError('Missing JWT token.'));
 	}
 
-	void verifyContextSymmetricJwtToken(token)
-		.then(({ connectInstallation, atlassianUserId }) => {
+	void verifyServerSymmetricJwtToken(req, token)
+		.then(({ connectInstallation }) => {
 			res.locals.connectInstallation = connectInstallation;
-			res.locals.atlassianUserId = atlassianUserId;
 			next();
 		})
-		.then(next);
+		.catch(next);
 };
 
 /**
  * @see https://developer.atlassian.com/cloud/jira/platform/understanding-jwt-for-connect-apps/#decoding-and-verifying-a-jwt-token
  */
-const verifyContextSymmetricJwtToken = async (token: string) => {
+const verifyServerSymmetricJwtToken = async (
+	request: Request,
+	token: string,
+) => {
 	try {
-		const tokenSigningAlgorithm: unknown = getAlgorithm(token);
+		const tokenSigningAlgorithm = getAlgorithm(token) as unknown;
 
 		if (!isEnumValueOf(SymmetricAlgorithm, tokenSigningAlgorithm)) {
 			throw new UnauthorizedError('Unsupported JWT signing algorithm.');
@@ -68,6 +68,8 @@ const verifyContextSymmetricJwtToken = async (token: string) => {
 
 		assertSchema(unverifiedClaims, CONNECT_JWT_TOKEN_CLAIMS_SCHEMA);
 
+		verifyIss(unverifiedClaims);
+
 		const connectInstallation =
 			await connectInstallationRepository.getByClientKey(
 				ensureString(unverifiedClaims.iss),
@@ -81,15 +83,14 @@ const verifyContextSymmetricJwtToken = async (token: string) => {
 
 		assertSchema(verifiedClaims, CONNECT_JWT_TOKEN_CLAIMS_SCHEMA);
 
-		verifyFixedQsh(verifiedClaims, CONTEXT_TOKEN_QSH);
+		verifyUrlBoundQsh(verifiedClaims, request);
 		verifyExp(verifiedClaims);
 
 		return {
 			connectInstallation,
-			atlassianUserId: ensureString(verifiedClaims.sub),
 		};
 	} catch (e) {
-		getLogger().warn(e, 'Failed to verify the context symmetric JWT token.');
+		getLogger().warn(e, 'Failed to verify the server symmetric JWT token.');
 
 		if (e instanceof UnauthorizedError) throw e;
